Convert error.c and add DragonFly base/mk-style Makefile
All libintl references need to be removed.
The libnvpair that is used is installed in /usr/local prefix so
we need to be careful not to accidently bring in local s/w.

BSD doesn't have "once" variation of thread key create so we
need to keep using the "MT unsafe" code. (at least unsafe for Solaris)