Rewrite URI validation
The URI validation uses regcmp() and regex() which doesn't exist on
DragonFly.  Moreever, it's doing a lot of extra work to split the
URI into components when all it's asked to do is check if the URL
is valid.

I used the regex listed in RFC 2396 Appendix B to verify the
given URI against it.  The only check needed after that is to ensure
the "path" part has a length greater than zero.

The only downside to this approach is that regcomp is run every
time the function is.  It's possible to do this once and store in
a static variable, but I'll leave this optimization for now.